requirements.txt
Flask>=2.3.0
gunicorn>=21.2.0
Flask-MySQLdb>=1.0.1
APScheduler>=3.10.4
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0

livekit-agents>=0.8.0
livekit-api>=0.6.0
livekit-plugins-openai>=0.6.0
livekit-plugins-cartesia>=0.6.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-silero>=0.6.0
livekit-plugins-noise-cancellation
